Move tanh to paddle.tensor.math test=develop
