+++ python/paddle/nn/functional/activation.py
@@ -23,6 +23,7 @@ from ...fluid.layers import soft_relu  #DEFINE_ALIAS
 from ...fluid.layers import swish  #DEFINE_ALIAS
 from ...fluid.layers import sigmoid  #DEFINE_ALIAS
 from ...fluid.layers import thresholded_relu  #DEFINE_ALIAS
+from ...tensor.math import tanh  #DEFINE_ALIAS
 
 __all__ = [
     'brelu',
@@ -210,47 +211,6 @@ def hardshrink(x, threshold=0.5, name=None):
         inputs={'X': x},
         outputs={'Out': out},
         attrs={'threshold': threshold})
-    return out
-
-
-def tanh(x, name=None):
-    """
-    tanh activation
-
-    .. math::
-
-        tanh(x) = (exp(x) - exp(-x))/(exp(x) + exp(-x))
-
-    Args:
-        x (Tensor): The input Tensor with data type float32, float16.
-        name (str, optional): Name for the operation (optional, default is None).
-            For more information, please refer to :ref:`api_guide_Name`.
-
-    Returns:
-        A Tensor with the same data type and shape as ``x`` .
-
-    Examples:
-
-        .. code-block:: python
-
-        import paddle
-        import paddle.nn.functional as F
-        import numpy as np
-
-        paddle.disable_static()
-
-        x = paddle.to_variable(np.array([-1, 0.5, 1]))
-        out = F.tanh(x) # [-0.761594, 0.462117, 0.761594]
-
-
-    """
-    if in_dygraph_mode():
-        return core.ops.tanh(x)
-
-    check_variable_and_dtype(x, 'x', ['float16', 'float32', 'float64'], 'tanh')
-    helper = LayerHelper('tanh', **locals())
-    out = helper.create_variable_for_type_inference(x.dtype)
-    helper.append_op(type='tanh', inputs={'X': x}, outputs={'Out': out})
     return out
